move func to postgres
